test init

+++ app/Console/Commands/InitCommand.js
@@ -112,14 +112,16 @@ export default class InitCommand extends Command {
           await mv(`${__dirname}/../../../data/vcc.db`, `${user.homedir}/.npm/vcc/data/vcc.db`, { mkdirp: true });
           fs.symlinkSync(`${user.homedir}/.npm/vcc/data/vcc.db`, `${__dirname}/../../../data/vcc.db`);
         } else {
-          const answers = await inquirer.prompt({ type: 'confirm', name: 'config', message: 'Database exitis - you overwrite ?', default: false });
-          if (answers.config) {
-            await rimraf(`${user.homedir}/.npm/vcc/data/vcc.db`);
-            await mv(`${__dirname}/../../../data/vcc.db`, `${user.homedir}/.npm/vcc/data/vcc.db`, { mkdirp: true });
-            fs.symlinkSync(`${user.homedir}/.npm/vcc/data/vcc.db`, `${__dirname}/../../../data/vcc.db`);
-          } else {
-            await rimraf(`${__dirname}/../../../data/vcc.db`);
-            fs.symlinkSync(`${user.homedir}/.npm/vcc/data/vcc.db`, `${__dirname}/../../../data/vcc.db`);
+          if ((await lstat(`${__dirname}/../../data/vcc.db`)).isSymbolicLink()) {
+            const answers = await inquirer.prompt({ type: 'confirm', name: 'config', message: 'Database exitis - you overwrite ?', default: false });
+            if (answers.config) {
+              await rimraf(`${user.homedir}/.npm/vcc/data/vcc.db`);
+              await mv(`${__dirname}/../../../data/vcc.db`, `${user.homedir}/.npm/vcc/data/vcc.db`, { mkdirp: true });
+              fs.symlinkSync(`${user.homedir}/.npm/vcc/data/vcc.db`, `${__dirname}/../../../data/vcc.db`);
+            } else {
+              await rimraf(`${__dirname}/../../../data/vcc.db`);
+              fs.symlinkSync(`${user.homedir}/.npm/vcc/data/vcc.db`, `${__dirname}/../../../data/vcc.db`);
+            }
           }
         }
